+++ src/composables/visibility.ts
@@ -46,12 +46,27 @@ function getVisibilityOptions(author: User, inReplyTo: Post | null): Visibility[
 export function useVisibility() {
   const { t } = useI18n({ useScope: "global" })
 
-  const VISIBILITY_MAP = {
-    [Visibility.Public]: t("post.visibility_public"),
-    [Visibility.Followers]: t("post.visibility_followers"),
-    [Visibility.Subscribers]: t("post.visibility_subscribers"),
-    [Visibility.Conversation]: t("post.visibility_conversation"),
-    [Visibility.Direct]: t("post.visibility_direct"),
+  const VISIBILITY_MAP: { [name in Visibility]: { name: string, description: string } } = {
+    [Visibility.Public]: {
+      name: t("post.visibility.public"),
+      description: t("post.visibility_description.visible_to_everyone"),
+    },
+    [Visibility.Followers]: {
+      name: t("post.visibility.followers"),
+      description: t("post.visibility_description.visible_to_followers"),
+    },
+    [Visibility.Subscribers]: {
+      name: t("post.visibility.subscribers"),
+      description: t("post.visibility_description.visible_to_premium_subscribers"),
+    },
+    [Visibility.Conversation]: {
+      name: t("post.visibility.conversation"),
+      description: t("post.visibility_description.visible_to_people_chosen_by_conversation_owner"),
+    },
+    [Visibility.Direct]: {
+      name: t("post.visibility.direct"),
+      description: t("post.visibility_description.visible_to_mentioned_users"),
+    },
   }
 
   return {