+++ src/components/VisibilityIcon.vue
@@ -1,7 +1,7 @@
 <template>
   <icon-globe v-if="visibility === 'public'"></icon-globe>
   <icon-users v-else-if="visibility === 'private'"></icon-users>
-  <icon-payment v-else-if="visibility === 'subscribers'"></icon-payment>
+  <icon-premium v-else-if="visibility === 'subscribers'"></icon-premium>
   <icon-lock v-else-if="visibility === 'conversation'"></icon-lock>
   <icon-envelope v-else-if="visibility === 'direct'"></icon-envelope>
 </template>
@@ -10,7 +10,7 @@
 import IconLock from "@/assets/feather/lock.svg?component"
 import IconGlobe from "@/assets/feather/globe.svg?component"
 import IconEnvelope from "@/assets/forkawesome/envelope-o.svg?component"
-import IconPayment from "@/assets/tabler/coin.svg?component"
+import IconPremium from "@/assets/extra-icons/spark.svg?component"
 import IconUsers from "@/assets/tabler/users.svg?component"
 
 defineProps<{