+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support RTL text in posts and profile descriptions.
+
 ## [3.8.0] - 2024-10-25
 
 ### Added