+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support RTL text in posts and profile descriptions.
 
+### Fixed
+
+- Try to fix FreeBSD build.
+
 ## [3.8.0] - 2024-10-25
 
 ### Added