+++ CHANGELOG.md
@@ -10,9 +10,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support RTL text in posts and profile descriptions.
 
+### Removed
+
+- Removed FreeBSD patch.
+
 ### Fixed
 
-- Try to fix FreeBSD build.
+- Fix FreeBSD build.
 
 ## [3.8.0] - 2024-10-25