+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support RTL text in posts and profile descriptions.
+- Allow followers-only self-replies to followers-only posts.
 
 ### Removed