+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Improved descriptions of visibility scopes.
 - Changed subscription icon.
+- Don't hide post menu after bookmarking.
 
 ### Removed