+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved descriptions of visibility scopes.
 - Changed subscription icon.
 - Don't hide post menu after bookmarking.
+- Move "Hide image" button to top right corner of image.
 
 ### Removed